fixed som of my mistakes in the accounts settings route and adjusted so the change settingss now shows the contributor or viewer settings and not the res=gistered user settings
added a logout route and adjusted the index rout to include it

added a new view layout to extend from when a user is logged in. this
new view adds the logout button when a user is logged in. adjusted the
relevent views so they extend layoutLoggedIn instead of layout.